+++ db/upgrade.php
@@ -261,5 +261,9 @@ function xmldb_enrol_autoenrol_upgrade($oldversion) {
         upgrade_plugin_savepoint(true, 2021112600, 'enrol', 'autoenrol');
     }
 
+    if ($oldversion < 2022031600) {
+        upgrade_plugin_savepoint(true, 2022031600, 'enrol', 'autoenrol');
+    }
+
     return true;
 }